power: supply: rk817: Fix battery capacity sanity check calculation
If the nvram variable is corrupted in a way that multiplying it by 1000 causes
an integer overflow and wraps around to a number within or below the sanity
check bounds, the check fails and capacity is reported incorrectly.

Change the sanity check calculation to not multiply the variable read from the
rk817 nvram, preventing the overflow and resulting in all corrupt capacity
values being detected.

Signed-off-by: Alexander Warnecke <awarnecke002@hotmail.com>