drivers: staging: bes2600: Enable automatic module loading
I think the new driver is stable enough (from what I've seen in the
PineTab channel).

If we don't let people use it, we'll never figure out if it's stable
enough.

Of course, some people still prefer using an external Wi-Fi adapter
so if they want, they can blacklist the "bes2600" module and continue
to using it.